Preparative refactor
