+++ app/src/engineering/java/com/hedvig/app/feature/insurance/MockInsuranceViewModel.kt
@@ -1,6 +1,7 @@
 package com.hedvig.app.feature.insurance
 
-import com.hedvig.app.feature.insurance.ui.InsuranceViewModel
+import com.hedvig.app.feature.insurance.ui.tab.InsuranceViewModel
+import com.hedvig.app.feature.insurance.ui.tab.items
 import com.hedvig.app.testdata.feature.insurance.INSURANCE_DATA_SWEDISH_APARTMENT
 
 class MockInsuranceViewModel : InsuranceViewModel() {
@@ -15,7 +16,7 @@ class MockInsuranceViewModel : InsuranceViewModel() {
             _data.value = ViewState.Error
             return
         }
-        _data.value = ViewState.Success(insuranceMockData)
+        _data.value = ViewState.Success(items(insuranceMockData))
     }
 
     companion object {
