+++ app/src/main/java/com/hedvig/app/ApplicationModule.kt
@@ -64,12 +64,13 @@ import com.hedvig.app.feature.home.ui.changeaddress.ChangeAddressViewModel
 import com.hedvig.app.feature.home.ui.changeaddress.ChangeAddressViewModelImpl
 import com.hedvig.app.feature.home.ui.changeaddress.GetAddressChangeStoryIdUseCase
 import com.hedvig.app.feature.home.ui.changeaddress.GetUpcomingAgreementUseCase
-import com.hedvig.app.feature.insurance.data.InsuranceRepository
+import com.hedvig.app.feature.insurance.data.GetContractsUseCase
 import com.hedvig.app.feature.insurance.service.InsuranceTracker
-import com.hedvig.app.feature.insurance.ui.InsuranceViewModel
-import com.hedvig.app.feature.insurance.ui.InsuranceViewModelImpl
 import com.hedvig.app.feature.insurance.ui.detail.ContractDetailViewModel
 import com.hedvig.app.feature.insurance.ui.detail.ContractDetailViewModelImpl
+import com.hedvig.app.feature.insurance.ui.tab.InsuranceViewModel
+import com.hedvig.app.feature.insurance.ui.tab.InsuranceViewModelImpl
+import com.hedvig.app.feature.insurance.ui.terminatedcontracts.TerminatedContractsViewModel
 import com.hedvig.app.feature.keygear.KeyGearTracker
 import com.hedvig.app.feature.keygear.KeyGearValuationViewModel
 import com.hedvig.app.feature.keygear.KeyGearValuationViewModelImpl
@@ -320,6 +321,7 @@ val viewModelModule = module {
             multiActionParams
         )
     }
+    viewModel { TerminatedContractsViewModel(get()) }
 }
 
 val choosePlanModule = module {
@@ -443,7 +445,6 @@ val repositoriesModule = module {
     single { ChatRepository(get(), get(), get()) }
     single { PayinStatusRepository(get()) }
     single { ClaimsRepository(get(), get()) }
-    single { InsuranceRepository(get(), get()) }
     single { ProfileRepository(get()) }
     single { RedeemReferralCodeRepository(get(), get()) }
     single { UserRepository(get()) }
@@ -529,6 +530,7 @@ val useCaseModule = module {
     single { ApproveQuotesUseCase(get(), get(), get()) }
     single { RefreshQuotesUseCase(get()) }
     single { LogoutUseCase(get(), get(), get(), get(), get(), get(), get(), get()) }
+    single { GetContractsUseCase(get(), get()) }
 }
 
 val cacheManagerModule = module {
