+++ app/src/main/java/com/hedvig/app/feature/loggedin/ui/LoggedInTabs.kt
@@ -5,7 +5,7 @@ import androidx.annotation.IdRes
 import androidx.fragment.app.Fragment
 import com.hedvig.app.R
 import com.hedvig.app.feature.home.ui.HomeFragment
-import com.hedvig.app.feature.insurance.ui.InsuranceFragment
+import com.hedvig.app.feature.insurance.ui.tab.InsuranceFragment
 import com.hedvig.app.feature.keygear.ui.tab.KeyGearFragment
 import com.hedvig.app.feature.profile.ui.tab.ProfileFragment
 import com.hedvig.app.feature.referrals.ui.tab.ReferralsFragment
