+++ app/src/main/res/layout/fragment_insurance.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.insurance.ui.InsuranceFragment">
+    tools:context=".feature.insurance.ui.tab.InsuranceFragment">
 
     <include
         android:id="@+id/loadSpinner"
